+++ src/clj_async_profiler/ui.clj
@@ -1,11 +1,19 @@
 (ns clj-async-profiler.ui
   (:require [clj-async-profiler.core :as core]
             [clj-async-profiler.wwws :as wwws :refer [redirect respond]]
+            [clj-async-profiler.render :as render]
+            [clj-async-profiler.results :as results]
             [clojure.java.io :as io]
             [clojure.string :as str])
-  (:import java.io.File))
+  (:import java.io.File
+           java.util.Date
+           java.time.format.DateTimeFormatter
+           java.time.ZoneId))
 
 (def ^:private ui-state (atom {}))
+
+(defmacro ^:private forj {:style/indent 1} [bindings body]
+  `(str/join (for ~bindings ~body)))
 
 (defn- profiler-controls []
   (let [^String status-msg (core/status)]
@@ -16,54 +24,81 @@
       (format "<form action=\"/start-profiler\">Event:&nbsp;
                <select name=\"event\">%s</select>&nbsp;
                <input type=\"submit\" value=\"Start profiler\"/></form>"
-              (str/join
-               (for [ev (core/list-event-types {:silent? true})]
-                 (format "<option value=%s>%s</option>" (name ev) (name ev))))))))
+              (forj [ev (remove #{:ClassName.methodName}
+                                (core/list-event-types {:silent? true}))]
+                (format "<option value=%s>%s</option>" (name ev) (name ev)))))))
 
-(defn- main-page
-  [root files]
-  (let [{:keys [show-raw-files]} @ui-state]
-   (format
-    "<html><head>
-<title>clj-async-profiler</title>
-<link rel='icon' href='favicon.png' type='image/x-icon'/>
-<style>
-body {
-margin: 1em auto;
-max-width: 40em;
-font: 1.1em/1.5 sans-serif;
-}
-a { text-decoration: none; }
-</style></head><body><div><big>clj-async-profiler</big>
-<small style=\"float: right\"><a href=\"/toggle-show-raw\">%s</a> | <a href=\"/clear-results\">Clear all results</a></small></div><hr>
-<small>%s</small>
-<hr><ul>%s</ul><hr></body></html>"
-    (if show-raw-files "Hide raw files" "Show raw files")
-    (profiler-controls)
-    (->> (for [^File f files
-               :let [fname (.getName f)
-                     sz (.length f)]]
-           (format "<li><a href='%s'>%s</a> <small><font color=\"gray\">(%s%s)</font></small></li>"
-                   (str root fname)
-                   fname
-                   (if-let [{:keys [samples]} (@@#'core/flamegraph-file->metadata f)]
-                     (if (< samples 1000)
-                       (str samples " samples, ")
-                       (format "%.1fk samples, " (/ samples 1e3)))
-                     "")
-                   (if (< sz 1024) (str sz " b") (format "%.1f Kb" (/ sz 1024.0)))))
-         str/join))))
+(def ^:private ^DateTimeFormatter date-formatter (DateTimeFormatter/ofPattern "MMM d, yyyy"))
+(def ^:private ^DateTimeFormatter time-formatter (DateTimeFormatter/ofPattern "HH:mm"))
+
+(defn- date->local-date [^Date date]
+  (.format date-formatter (.atZone (.toInstant date) (ZoneId/systemDefault))))
+
+(defn- date->local-time [^Date date]
+  (.format time-formatter (.atZone (.toInstant date) (ZoneId/systemDefault))))
+
+(defn- row [tag tr-classes & cells]
+  (let [tag (name tag)
+        tds (forj [c cells]
+              (format "<%s>%s</%s>" tag (or c "") tag))]
+    (format "<tr class='%s'>%s</tr>" tr-classes tds)))
+
+(defn- format-size [sz samples]
+  (let [sz (if (< sz 1024) (str sz " b") (format "%.1f Kb" (/ sz 1024.0)))
+        samples (cond (nil? samples) nil
+                      (< samples 1000) (str "<br>" samples " samples")
+                      :else (format "<br>%.1fk samples" (/ samples 1e3)))]
+    (cond-> sz samples (str samples))))
+
+(defn- file-table [root files]
+  (let [files (sort-by #(.lastModified ^File %) > files) ;; Newer on top
+        parsed-files (for [^File f files
+                           :let [info (results/parse-results-filename (.getName f))]]
+                       [f (update info :date #(or % (Date. (.lastModified f))))])
+        grouped (group-by #(date->local-date (:date (second %))) parsed-files)]
+    (format
+     "<table><thead>%s</thead><tbody>%s</tbody></table>"
+     (row :th "heading"
+          "Date/time" "ID" "File" "Event" "Size" "Extras" "Actions")
+     (forj [[local-date day-files] grouped]
+       (str (row :td "heading"
+                 (str "" local-date "") "" "" "" "" "" "")
+            (forj [[^File f {:keys [id date kind event]}] day-files
+                   :let [filename (.getName f)
+                         ^File stacks (results/find-stacks-file-by-flamegraph-file f)
+                         {:keys [id1 id2 ^File stacks1 ^File stacks2]}
+                         (@results/file->metadata f)]]
+              (row :td "data"
+                   (date->local-time date)
+                   (when (and id stacks
+                              (= (:stacks-file (results/find-profile id))
+                                 stacks))
+                     (format "%02d" id))
+                   (format "<a href='%s'>%s</a>"
+                           (str root filename)
+                           (or kind filename))
+                   (some-> event name)
+                   (str "<small>" (format-size (.length f) (:samples (@results/file->metadata f))) "</small>")
+                   (cond id1 (format "<a href='%s'>%02d</a> vs <a href='%s'>%02d</a>"
+                                     (str root (.getName ^File stacks1))
+                                     id1
+                                     (str root (.getName ^File stacks2))
+                                     id2)
+                         stacks (format "<a href='%s'>raw</a>"
+                                        (str root (.getName stacks))))
+                   "")))))))
+
+(defn- main-page [root files]
+  (->> {:profiler-controls (profiler-controls)
+        :file-table (file-table root files)}
+       (render/render-template (slurp (io/resource "ui/index.html")))))
 
 (defn- handler [{:keys [path params]} base]
   (try
     (let [files (->> (.listFiles (io/file base))
                      (remove #(.isDirectory ^File %))
                      sort)]
-      (cond (= path "/toggle-show-raw")
-            (do (swap! ui-state update :show-raw-files not)
-                (redirect "/"))
-
-            (= path "/start-profiler")
+      (cond (= path "/start-profiler")
             (do (core/start {:event (keyword (params "event"))})
                 (redirect "/"))
 
@@ -76,10 +111,7 @@ a { text-decoration: none; }
                 (redirect "/"))
 
             (= path "/")
-            (let [files (if (:show-raw-files @ui-state)
-                          files
-                          (remove #(= (wwws/get-extension (str %)) "txt") files))]
-              {:body (main-page path files)})
+            {:body (main-page path files)}
 
             (= path "/favicon.png")
             {:body (io/resource "favicon.png")