+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### master (unreleased)
 
+- New index page design.
+
 ### 1.3.3 (2024-10-04)
 
 - [async-profiler#932](https://github.com/async-profiler/async-profiler/issues/923): Fix SIGSEVG crash on JDK23.